Split mono.rs into mono/ directory with discovery, specialization, rewrite, propagation, utils modules
